Reorganize files
To keep things tidy, the src/lib folder was split up into a folder
containing grammar info (for instance, with verb conjugations to come)
and a helpers folder for things like the shuffling function.

Furthermore, all the existing components were split into a shared folder
and one for guessing Latin nouns, which will be mirrored by guessing
Barbarian translations, verbs and so on.