Improve handling of words in app state
Previously, words were initially loaded from the API and removed from
the state as they were presented to the user for studying.
Now, a counter keeps track of where in the list of words the student is,
preserving the list of all the studied words for that session. This is
useful as something can be done with this list.