Reorganize components dir
Components are now better organized. The root folder contains *only* the
main App component. The components it might call as different versions
of the page are under Pages, rather than the root or Words like before.

Shared contains a 'private' folder, for components that are not exposed
to others using the shared elements. And playing with the exports in
Words allows for a better structure of Drill, which decides what type of
drill to show the user.